adjust imblances
