+++ tokens/Cargo.toml
@@ -21,6 +21,8 @@ orml-traits = { path = "../traits", version = "0.1.3-dev", default-features = fa
 
 [dev-dependencies]
 sp-core = { version = "2.0.0-rc6", default-features = false }
+pallet-treasury = "2.0.0-rc6"
+pallet-elections-phragmen = "2.0.0-rc6"
 
 clear_on_drop = { version = "0.2.4", features = ["no_cc"] }	# https://github.com/paritytech/substrate/issues/4179
